Comments: fall back to site post link until deep reader links for jetpack are supported
