+++ client/blocks/comment-detail/comment-detail-post.jsx
@@ -2,8 +2,8 @@
  * External dependencies
  */
 import React from 'react';
-import { connect } from 'react-redux';
 import { localize } from 'i18n-calypso';
+import { noop } from 'lodash';
 
 /**
  * Internal dependencies
@@ -11,11 +11,6 @@ import { localize } from 'i18n-calypso';
 import Emojify from 'components/emojify';
 import Gravatar from 'components/gravatar';
 import SiteIcon from 'blocks/site-icon';
-import {
-	bumpStat,
-	composeAnalytics,
-	recordTracksEvent,
-} from 'state/analytics/actions';
 
 export const CommentDetailPost = ( {
 	commentId,
@@ -25,8 +20,7 @@ export const CommentDetailPost = ( {
 	postAuthorDisplayName,
 	postTitle,
 	postUrl,
-	recordReaderArticleOpened,
-	recordReaderCommentOpened,
+	onClick = noop,
 	siteId,
 	translate,
 } ) => {
@@ -50,7 +44,7 @@ export const CommentDetailPost = ( {
 							</Emojify>
 						</span>
 					}
-					<a href={ `${ postUrl }#comment-${ commentId }` } onClick={ recordReaderCommentOpened }>
+					<a href={ `${ postUrl }#comment-${ commentId }` } onClick={ onClick }>
 						<Emojify>
 							{ parentCommentContent }
 						</Emojify>
@@ -71,7 +65,7 @@ export const CommentDetailPost = ( {
 						</Emojify>
 					</span>
 				}
-				<a href={ postUrl } onClick={ recordReaderArticleOpened }>
+				<a href={ postUrl } onClick={ onClick }>
 					<Emojify>
 						{ postTitle || translate( 'Untitled' ) }
 					</Emojify>
@@ -81,15 +75,4 @@ export const CommentDetailPost = ( {
 	);
 };
 
-const mapDispatchToProps = dispatch => ( {
-	recordReaderArticleOpened: () => dispatch( composeAnalytics(
-		recordTracksEvent( 'calypso_comment_management_article_opened' ),
-		bumpStat( 'calypso_comment_management', 'article_opened' )
-	) ),
-	recordReaderCommentOpened: () => dispatch( composeAnalytics(
-		recordTracksEvent( 'calypso_comment_management_comment_opened' ),
-		bumpStat( 'calypso_comment_management', 'comment_opened' )
-	) ),
-} );
-
-export default connect( null, mapDispatchToProps )( localize( CommentDetailPost ) );
+export default localize( CommentDetailPost );
